Elevated to protected the registerContact method
The level to protected of the registerContact method is increased in order to manage the registration procedure from child controllers. At the same time the boolean return is added if the contact has been correctly registered and the redirecting method is separated from the registration so that it can be optional for the child controllers.